feat: 适配 RK 原生 HDMI IN 适配采集

+++ src/web/handlers/mod.rs
@@ -1001,7 +1001,7 @@ pub struct VideoFormat {
 pub struct VideoResolution {
     pub width: u32,
     pub height: u32,
-    pub fps: Vec<u32>,
+    pub fps: Vec<f64>,
 }
 
 #[derive(Serialize)]
